move ExpectedAssignment to assignment package
